reoganize carousel
