--- src/components/Carousel/CarouselControls.tsx
@@ -1,41 +1,0 @@
-// components/CarouselControls.tsx
-import classNames from "classnames";
-import {Props} from '@/types'
-
-const CarouselControls = (props: Props) => {
-  return (
-    <div className="flex justify-end gap-2 ">
-      <button
-        onClick={() => {
-          if (props.canScrollPrev) {
-            props.onPrev();
-          }
-        }}
-        disabled={!props.canScrollPrev}
-        className={classNames({
-          "px-4 py-2 text-white rounded-md": true,
-          "bg-indigo-200": !props.canScrollPrev,
-          "bg-indigo-400": props.canScrollPrev,
-        })}
-      >
-        Prev
-      </button>
-      <button
-        onClick={() => {
-          if (props.canScrollNext) {
-            props.onNext();
-          }
-        }}
-        disabled={!props.canScrollNext}
-        className={classNames({
-          "px-4 py-2 text-white rounded-md": true,
-          "bg-indigo-200": !props.canScrollNext,
-          "bg-indigo-400": props.canScrollNext,
-        })}
-      >
-        Next
-      </button>
-    </div>
-  );
-};
-export default CarouselControls;
